requirements.txt
Django==1.4.1
django-tastypie==0.9.11
mimeparse==0.1.3
python-dateutil==1.5
vboxapi==1.0
wsgiref==0.1.2
